requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bs4==0.0.2
chromadb==0.6.3
cryptography==44.0.2
deep-translator==1.11.4
deepface==0.0.93
docutils==0.21.2
filelock==3.17.0
googletrans==4.0.2
gTTS==2.5.4
hid-converge==2.0.1.1
HTMLParser==0.0.2
httpsproxy_urllib2==1.0
import-java==0.6
importlib_metadata==8.5.0
Jinja2==3.1.6
langchain==0.3.21
langchain-chroma==0.2.2
langchain-community==0.3.20
langchain-core==0.3.50
langchain-google-genai==2.1.2
openai==1.68.2
opencv-python==4.11.0.86
pandas==2.2.3
pillow==10.2.0
pip==25.0.1
PyConfigParser==1.0.5
pygame==2.6.1
pymongo==4.11.3
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python-ntlm==1.1.0
railroad==0.5.0
redis==5.2.1
requests==2.32.3
spacy==3.7.5
SpeechRecognition==3.14.1
Sphinx==8.3.0
streamlit==1.43.1
tornado==6.4.2
transformers==4.49.0

# WARNING(pigar): some manual fixes might be required as pigar has detected duplicate requirements for the same import name (possibly for different submodules).
# WARNING(pigar): the following duplicate requirements are for the import name: keyring
keyring==25.6.0
pyKeyring==0.6
# WARNING(pigar): the following duplicate requirements are for the import name: TTS
coqui-tts==0.26.0
TTS==0.22.0
